wlsy v1.0

+++ targets/CH32V20X/src/app/wlsy/output.hpp
@@ -14,7 +14,7 @@ public:
     CaptureChannelExti & cap;
     real_t dur_ms;//ms
     
-    static constexpr uint8_t poles = 9;
+    static constexpr uint poles = 9;
     static constexpr real_t rad_delta = real_t(TAU / poles);
     static constexpr real_t radius = real_t(0.05);
 public:
